Revert "refactor(TripDetails): Extract defaults.js and messages.js rendering code."
This reverts commit 1e74af75def2a8325968a35b46f2d139374ea854.
