+++ packages/trip-details/src/index.js
@@ -2,18 +2,135 @@ import coreUtils from "@opentripplanner/core-utils";
 import moment from "moment";
 import PropTypes from "prop-types";
 import React from "react";
+import { FormattedMessage } from "react-intl";
 import { CalendarAlt, Heartbeat, MoneyBillAlt } from "styled-icons/fa-solid";
 
-import {
-  DefaultDepart,
-  DefaultTransitFare,
-  DefaultTNCFare,
-  DefaultCaloriesBurned,
-  DefaultCaloriesBurnedDescription
-} from "./defaults";
-import Message from "./message";
 import * as Styled from "./styled";
 import TripDetail from "./trip-detail";
+
+/**
+ * Default rendering for the departure date/time line
+ * if no other corresponding message is provided.
+ */
+function renderDefaultDepart(itinerary, longDateFormat, timeOptions) {
+  const date = moment(itinerary.startTime);
+  return (
+    <>
+      Depart <b>{date.format(longDateFormat)}</b>
+      <> at </>
+      <b>{coreUtils.time.formatTime(itinerary.startTime, timeOptions)}</b>
+    </>
+  );
+}
+
+/**
+ * Default rendering for the transit fare line
+ * if no other corresponding message is provided.
+ */
+function renderDefaultTransitFare(fareResult) {
+  const { centsToString, transitFare } = fareResult;
+  return (
+    <>
+      Transit Fare: <b>{centsToString(transitFare)}</b>
+    </>
+  );
+}
+
+/**
+ * Default rendering for the TNC fare line
+ * if no other corresponding message is provided.
+ */
+function renderDefaultTNCFare(itinerary, fareResult) {
+  const { dollarsToString, maxTNCFare, minTNCFare } = fareResult;
+  let companies;
+  itinerary.legs.forEach(leg => {
+    if (leg.tncData) {
+      companies = leg.tncData.company;
+    }
+  });
+  return (
+    <>
+      <Styled.TNCFareCompanies>
+        {companies.toLowerCase()}
+      </Styled.TNCFareCompanies>
+      {" fare: "}
+      <b>
+        {dollarsToString(minTNCFare)} - {dollarsToString(maxTNCFare)}
+      </b>
+    </>
+  );
+}
+
+/**
+ * Default rendering for the calories burned line
+ * if no other corresponding message is provided.
+ */
+function renderDefaultCaloriesBurned(caloriesBurned) {
+  return (
+    <>
+      Calories Burned: <b>{Math.round(caloriesBurned)}</b>
+    </>
+  );
+}
+
+/**
+ * Default rendering for the calories description
+ * if no other corresponding message is provided.
+ */
+function renderDefaultCaloriesBurnedDescription(walkDuration, bikeDuration) {
+  return (
+    <>
+      Calories burned is based on{" "}
+      <b>{Math.round(walkDuration / 60)} minute(s)</b> spent walking and{" "}
+      <b>{Math.round(bikeDuration / 60)} minute(s)</b> spent biking during this
+      trip. Adapted from{" "}
+      <a
+        href="https://health.gov/dietaryguidelines/dga2005/document/html/chapter3.htm#table4"
+        rel="noopener noreferrer"
+        target="_blank"
+      >
+        Dietary Guidelines for Americans 2005, page 16, Table 4
+      </a>
+      .
+    </>
+  );
+}
+
+/**
+ * Format text bold (used with FormattedMessage).
+ */
+// TODO: Find a better place for this component.
+function BoldText(chunks) {
+  return <b>{chunks}</b>;
+}
+
+/**
+ * Renders a message depending on parameters.
+ */
+// TODO: Find a better place for this component.
+function Message({ defaultContent, localized, messageId, values }) {
+  return (
+    (localized && (
+      <FormattedMessage id={messageId} values={{ b: BoldText, ...values }} />
+    )) ||
+    messageId ||
+    defaultContent
+  );
+}
+
+Message.propTypes = {
+  defaultContent: PropTypes.elementType,
+  localized: PropTypes.bool,
+  messageId: PropTypes.string,
+  values: PropTypes.shape({})
+};
+
+Message.defaultProps = {
+  defaultContent: null,
+  localized: false,
+  messageId: null,
+  values: null
+};
 
 export default function TripDetails({
   className,
@@ -46,7 +163,7 @@ export default function TripDetails({
         {transitFare && (
           <Styled.TransitFare>
             <Message
-              defaultContent={<DefaultTransitFare fareResult={fareResult} />}
+              defaultContent={renderDefaultTransitFare(fareResult)}
               localized={localized}
               messageId={messages.transitFare}
               values={{ transitFare }}
@@ -57,9 +174,7 @@ export default function TripDetails({
           <Styled.TNCFare>
             <br />
             <Message
-              defaultContent={
-                <DefaultTNCFare fareResult={fareResult} itinerary={itinerary} />
-              }
+              defaultContent={renderDefaultTNCFare(itinerary, fareResult)}
               localized={localized}
               messageId={messages.tncFare}
               values={{ companies, maxTNCFare, minTNCFare }}
@@ -96,13 +211,11 @@ export default function TripDetails({
           summary={
             <Styled.Timing>
               <Message
-                defaultContent={
-                  <DefaultDepart
-                    itinerary={itinerary}
-                    longDateFormat={longDateFormat}
-                    timeOptions={timeOptions}
-                  />
-                }
+                defaultContent={renderDefaultDepart(
+                  itinerary,
+                  longDateFormat,
+                  timeOptions
+                )}
                 localized={localized}
                 messageId={messages.depart}
                 values={{ departDate: moment(itinerary.startTime) }}
@@ -130,9 +243,7 @@ export default function TripDetails({
             summary={
               <Styled.CaloriesSummary>
                 <Message
-                  defaultContent={
-                    <DefaultCaloriesBurned caloriesBurned={caloriesBurned} />
-                  }
+                  defaultContent={renderDefaultCaloriesBurned(caloriesBurned)}
                   localized={localized}
                   messageId={messages.caloriesBurned}
                   values={{ caloriesBurned }}
@@ -143,12 +254,10 @@ export default function TripDetails({
               messages.caloriesBurnedDescription && (
                 <Styled.CaloriesDescription>
                   <Message
-                    defaultContent={
-                      <DefaultCaloriesBurnedDescription
-                        bikeDuration={bikeDuration}
-                        walkDuration={walkDuration}
-                      />
-                    }
+                    defaultContent={renderDefaultCaloriesBurnedDescription(
+                      walkDuration,
+                      bikeDuration
+                    )}
                     localized={localized}
                     messageId={messages.caloriesBurnedDescription}
                     values={{ bikeDuration, caloriesBurned, walkDuration }}
